Puck OSC Integration
puckaccel.js should be installed via the Web Bluetooth IDE or via espruino command-line utility.  Once installed, the puck will stream accelerometer and gyroscope data at 20 FPS to the bluetooth serial connection.  We can connect to that via espruino and pipe it to a simple python script to convert the data and send it via OSC to a local Chromatik instance.  It expects a couple of Knobs modulators configured with particular names.  An example setup is in puck.lxp.

espruino -p d7:98:e7:42:92:c4 | python puckosc.py