Shader support.
Basic shader implementation.  Note that when Chromatik imports the jar-with-dependencies.jar it will experience an error and show as red under 'Packages' but the shaders still work.  I tried some proguard rules but they are so broad that while keeping more stuff to get the trimmed jar to work it eventually included stuff that triggers the Chromatik error. The jogl native libraries (from natives/) needs to be put somewhere in the path.